feat: Implement kernel storage abstraction layer
Closes #118

+++ packages/kernel/src/kernel-store.ts
@@ -49,13 +49,16 @@ export function makeKernelStore(kv: KVStore) {
    * touch persistent storage if the value hasn't ever been read of if it is
    * modified; otherwise we can service read requests from memory.
    *
+   * IMPORTANT NOTE: in order for the cache to work, all subsequent accesses to
+   * the value MUST be made through a common stored value object.
+   *
    * @param key - A key string that identifies the value.
-   * @param init - A initial setting if the indicated value is not present.
+   * @param init - If provided, an initial setting if the stored entity does not exist.
    * @returns An object for interacting with the value.
    */
-  function makeCachedStoredValue(key: string, init: string): StoredValue {
-    let value: string | undefined;
-    if (kv.get(key) === undefined && init !== undefined) {
+  function provideCachedStoredValue(key: string, init?: string): StoredValue {
+    let value: string | undefined = kv.get(key);
+    if (value === undefined && init !== undefined) {
       kv.set(key, init);
       value = init;
     }
@@ -75,14 +78,14 @@ export function makeKernelStore(kv: KVStore) {
   }
 
   /**
-   * Provide a stored value object that is backed soley by persistent storage.
+   * Provide a stored value object that is kept in persistent storage without caching.
    *
    * @param key - A key string that identifies the value.
-   * @param init - A initial setting if the indicated value is not present.
+   * @param init - If provided, an initial setting if the stored entity does not exist.
    * @returns An object for interacting with the value.
    */
-  function makeRawStoredValue(key: string, init: string): StoredValue {
-    if (kv.get(key) === undefined) {
+  function provideRawStoredValue(key: string, init?: string): StoredValue {
+    if (kv.get(key) === undefined && init !== undefined) {
       kv.set(key, init);
     }
     return harden({
@@ -107,39 +110,64 @@ export function makeKernelStore(kv: KVStore) {
    */
   function incCounter(value: StoredValue): string {
     const current = value.get();
-    const next = `{Number(current) + 1}`;
+    const next = `${Number(current) + 1}`;
     value.set(next);
     return current as string;
   }
 
   /**
-   * Create a new persistently stored message queue.
+   * Create a new (empty) persistently stored message queue.
    *
-   * @param queueName - The name for the new queue (must be unique among queues).
+   * @param queueName - The name for the queue (must be unique among queues).
+   * @param cached - Optional flag: set to true if the queue should cache its
+   * @returns An object for interacting with the new queue.
+   */
+  function createStoredMessageQueue(
+    queueName: string,
+    cached: boolean = false,
+  ): StoredMessageQueue {
+    const qk = `queue.${queueName}`;
+    kv.set(`${qk}.head`, '1');
+    kv.set(`${qk}.tail`, '1');
+    return provideStoredMessageQueue(queueName, cached);
+  }
+
+  /**
+   * Produce an object to access a persistently stored message queue.
+   *
+   * @param queueName - The name for the queue (must be unique among queues).
    * @param cached - Optional flag: set to true if the queue should cache its
    * limit indices in memory (only do this if the queue is going to be accessed or
    * checked frequently).
    * @returns An object for interacting with the queue.
    */
-  function makeStoredMessageQueue(
+  function provideStoredMessageQueue(
     queueName: string,
     cached: boolean = false,
   ): StoredMessageQueue {
     const qk = `queue.${queueName}`;
-    // Note: cached==true ==> caches only the head & tail indices, NOT the messages themselves
-    const makeValue = cached ? makeCachedStoredValue : makeRawStoredValue;
-    const head = makeValue(`${qk}.head`, '1');
-    const tail = makeValue(`${qk}.tail`, '1');
+    // Note: cached=true ==> caches only the head & tail indices, NOT the messages themselves
+    const provideValue = cached
+      ? provideCachedStoredValue
+      : provideRawStoredValue;
+    const head = provideValue(`${qk}.head`);
+    const tail = provideValue(`${qk}.tail`);
     return {
       enqueue(message: Message): void {
+        if (head.get() === undefined) {
+          throw Error(`enqueue into deleted queue ${queueName}`);
+        }
         const entryPos = incCounter(head);
         kv.set(`${qk}.${entryPos}`, JSON.stringify(message));
       },
       dequeue(): Message | undefined {
         const headPos = head.get();
+        if (headPos === undefined) {
+          return undefined;
+        }
         const tailPos = tail.get();
         if (tailPos !== headPos) {
-          const entry = kv.get(`{qk}.${tailPos}`);
+          const entry = kv.get(`${qk}.${tailPos}`);
           kv.delete(`${qk}.${tailPos}`);
           incCounter(tail);
           return JSON.parse(entry) as Message;
@@ -148,19 +176,21 @@ export function makeKernelStore(kv: KVStore) {
       },
       delete(): void {
         const headPos = head.get();
-        let tailPos = tail.get();
-        while (tailPos !== headPos) {
-          kv.delete(`${qk}.${tailPos}`);
-          tailPos = `${Number(tailPos) + 1}`;
+        if (headPos !== undefined) {
+          let tailPos = tail.get();
+          while (tailPos !== headPos) {
+            kv.delete(`${qk}.${tailPos}`);
+            tailPos = `${Number(tailPos) + 1}`;
+          }
+          head.delete();
+          tail.delete();
         }
-        head.delete();
-        tail.delete();
       },
     };
   }
 
   /** The kernel's run queue. */
-  const runQueue = makeStoredMessageQueue('run', true);
+  const runQueue = createStoredMessageQueue('run', true);
 
   /**
    * Append a message to the kernel's run queue.
@@ -182,7 +212,7 @@ export function makeKernelStore(kv: KVStore) {
   }
 
   /** Counter for allocating VatIDs */
-  const nextVatId = makeCachedStoredValue('nextVatId', '1');
+  const nextVatId = provideCachedStoredValue('nextVatId', '1');
   /**
    * Obtain an ID for a new vat.
    *
@@ -193,7 +223,7 @@ export function makeKernelStore(kv: KVStore) {
   }
 
   /** Counter for allocating RemoteIDs */
-  const nextRemoteId = makeCachedStoredValue('nextRemoteId', '1');
+  const nextRemoteId = provideCachedStoredValue('nextRemoteId', '1');
   /**
    * Obtain an ID for a new remote connection.
    *
@@ -204,7 +234,7 @@ export function makeKernelStore(kv: KVStore) {
   }
 
   /** Counter for allocating kernel object IDs */
-  const nextObjectId = makeCachedStoredValue('nextObjectId', '1');
+  const nextObjectId = provideCachedStoredValue('nextObjectId', '1');
   /**
    * Obtain a KRef for the next unallocated kernel object.
    *
@@ -212,6 +242,52 @@ export function makeKernelStore(kv: KVStore) {
    */
   function getNextObjectId(): KRef {
     return `ko${incCounter(nextObjectId)}`;
+  }
+
+  /**
+   * Generate the storage key for a kernel entity's reference count.
+   *
+   * @param kref - The KRef of interest.
+   * @returns the key to store the indicated reference count at.
+   */
+  function refCtKey(kref: KRef): string {
+    return `${kref}.rc`;
+  }
+
+  /**
+   * Get a kernel entity's reference count.
+   *
+   * @param kref - The KRef of interest.
+   * @returns the reference count of the indicated kernel entity.
+   */
+  function getRefCt(kref: KRef): number {
+    return Number(kv.get(refCtKey(kref)));
+  }
+
+  /**
+   * Increment a kernel entity's reference count.
+   *
+   * @param kref - The KRef of the entity to increment the ref count of.
+   * @returns the new reference count after incrementing.
+   */
+  function incRefCt(kref: KRef): number {
+    const key = refCtKey(kref);
+    const newCount = Number(kv.get(key)) + 1;
+    kv.set(key, `${newCount}`);
+    return newCount;
+  }
+
+  /**
+   * Decrement a kernel entity's reference count.
+   *
+   * @param kref - The KRef of the entity to decrement the ref count of.
+   * @returns the new reference count after decrementing.
+   */
+  function decRefCt(kref: KRef): number {
+    const key = refCtKey(kref);
+    const newCount = Number(kv.get(key)) - 1;
+    kv.set(key, `${newCount}`);
+    return newCount;
   }
 
   /**
@@ -224,9 +300,10 @@ export function makeKernelStore(kv: KVStore) {
    * kernel object itself.
    */
   function initKernelObject(owner: EndpointId): [KRef, KernelObject] {
-    const kobj = { owner, reachableCount: 1, recognizableCount: 1 };
+    const kobj = { owner };
     const koid = getNextObjectId();
     kv.set(koid, JSON.stringify(kobj));
+    kv.set(refCtKey(koid), '1');
     return [koid, kobj];
   }
 
@@ -251,10 +328,11 @@ export function makeKernelStore(kv: KVStore) {
    */
   function deleteKernelObject(koid: KRef): void {
     kv.delete(koid);
+    kv.delete(refCtKey(koid));
   }
 
   /** Counter for allocating kernel promise IDs */
-  const nextPromiseId = makeCachedStoredValue('nextPromiseId', '1');
+  const nextPromiseId = provideCachedStoredValue('nextPromiseId', '1');
   /**
    * Obtain a KRef for the next unallocated kernel promise.
    *
@@ -277,13 +355,23 @@ export function makeKernelStore(kv: KVStore) {
     const kpr: KernelPromise = {
       decider,
       state: 'unresolved',
-      referenceCount: 1,
       value: undefined,
     };
     const kpid = getNextPromiseId();
-    makeStoredMessageQueue(`${kpid}.q`);
+    createStoredMessageQueue(kpid, false);
     kv.set(kpid, JSON.stringify(kpr));
+    kv.set(refCtKey(kpid), '1');
     return [kpid, kpr];
+  }
+
+  /**
+   * Append a message to a promise's message queue.
+   *
+   * @param kpid - The KRef of the promise to enqueue on.
+   * @param message - The message to enqueue.
+   */
+  function enqueuePromiseMessage(kpid: KRef, message: Message): void {
+    provideStoredMessageQueue(kpid, false).enqueue(message);
   }
 
   /**
@@ -308,7 +396,7 @@ export function makeKernelStore(kv: KVStore) {
    */
   function getKernelPromiseMessageQueue(kpid: KRef): Message[] {
     const result: Message[] = [];
-    const queue = makeStoredMessageQueue(`${kpid}.q`);
+    const queue = provideStoredMessageQueue(kpid, false);
     for (;;) {
       const message = queue.dequeue();
       if (message) {
@@ -326,8 +414,8 @@ export function makeKernelStore(kv: KVStore) {
    */
   function deleteKernelPromise(kpid: KRef): void {
     kv.delete(kpid);
-    const queue = makeStoredMessageQueue(`${kpid}.q`);
-    queue.delete();
+    kv.delete(refCtKey(kpid));
+    provideStoredMessageQueue(kpid).delete();
   }
 
   /**
@@ -355,11 +443,11 @@ export function makeKernelStore(kv: KVStore) {
   }
 
   /**
-   * Add an entry to a endpoints c-list, creating a new bidirectional mapping
+   * Add an entry to an endpoint's c-list, creating a new bidirectional mapping
    * between an ERef belonging to the endpoint and a KRef belonging to the
    * kernel.
    *
-   * @param endpointId - The endpoint whose c-list is to be updated.
+   * @param endpointId - The endpoint whose c-list is to be added to.
    * @param kref - The KRef.
    * @param eref - The ERef.
    */
@@ -368,21 +456,69 @@ export function makeKernelStore(kv: KVStore) {
     kv.set(`cle.${endpointId}.${eref}`, kref);
   }
 
+  /**
+   * Remove an entry from an endpoint's c-list.
+   *
+   * @param endpointId - The endpoint whose c-list entry is to be removed.
+   * @param kref - The KRef.
+   * @param eref - The ERef.
+   */
+  function deleteClistEntry(
+    endpointId: EndpointId,
+    kref: KRef,
+    eref: ERef,
+  ): void {
+    kv.delete(`clk.${endpointId}.${kref}`);
+    kv.delete(`cle.${endpointId}.${eref}`);
+  }
+
+  /**
+   * Remove an entry from an endpoint's c-list given an eref.
+   *
+   * @param endpointId - The endpoint whose c-list entry is to be removed.
+   * @param eref - The ERef.
+   */
+  function forgetEref(endpointId: EndpointId, eref: ERef): void {
+    const kref = erefToKref(endpointId, eref);
+    if (kref) {
+      deleteClistEntry(endpointId, kref, eref);
+    }
+  }
+
+  /**
+   * Remove an entry from an endpoint's c-list given a kref.
+   *
+   * @param endpointId - The endpoint whose c-list entry is to be removed.
+   * @param kref - The Kref.
+   */
+  function forgetKref(endpointId: EndpointId, kref: KRef): void {
+    const eref = krefToEref(endpointId, kref);
+    if (eref) {
+      deleteClistEntry(endpointId, kref, eref);
+    }
+  }
+
   return harden({
     enqueueRun,
     dequeueRun,
     getNextVatId,
     getNextRemoteId,
+    getRefCt,
+    incRefCt,
+    decRefCt,
     initKernelObject,
     getKernelObject,
     deleteKernelObject,
     initKernelPromise,
     getKernelPromise,
+    enqueuePromiseMessage,
     getKernelPromiseMessageQueue,
     deleteKernelPromise,
     erefToKref,
     krefToEref,
     addClistEntry,
+    forgetEref,
+    forgetKref,
     kv,
   });
 }
